Remove GLI, add API related casting if GFX_EXPOSE_APIS is defined

+++ src/shaders/ext/impl/sampling.inl
@@ -10,15 +10,15 @@ vec3 sample_cosine_hemisphere(vec2 uv)
     float y = r * sin(theta);
 
     // Project point up to the unit sphere
-    float z = sqrt(max(0.f, 1 - x * x - y * y));
+    float z = float(sqrt(max(0.f, 1 - x * x - y * y)));
     return vec3(x, y, z);
 }
 
 vec3 sample_uniform_sphere(vec2 uv)
 {
-	float cosTheta = 2 * uv.x - 1;
+	float cosTheta = float(2 * uv.x - 1);
 	float phi = uv.y * TAU;
-	float sinTheta = sqrt(max(0, 1 - cosTheta * cosTheta));
+	float sinTheta = float(sqrt(max(0.f, 1 - cosTheta * cosTheta)));
 	return vec3(sinTheta * cos(phi), sinTheta * sin(phi), cosTheta);
 }
